Replace two JavaCC lexers with a single JFlex one
