+++ instrumentation/jdbc/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jdbc/JdbcUtils.java
@@ -8,9 +8,7 @@ package io.opentelemetry.javaagent.instrumentation.jdbc;
 import static io.opentelemetry.javaagent.instrumentation.api.db.QueryNormalizationConfig.isQueryNormalizationEnabled;
 
 import io.opentelemetry.javaagent.instrumentation.api.db.SqlStatementInfo;
-import io.opentelemetry.javaagent.instrumentation.api.db.sanitizer.ParseException;
 import io.opentelemetry.javaagent.instrumentation.api.db.sanitizer.SqlSanitizer;
-import io.opentelemetry.javaagent.instrumentation.api.db.sanitizer.SqlStatementInfoExtractor;
 import java.lang.reflect.Field;
 import java.sql.Connection;
 import java.sql.Statement;
@@ -69,25 +67,10 @@ public abstract class JdbcUtils {
     return connection;
   }
 
-  /** Returns null if the sql could not be sanitized for any reason. */
-  public static String sanitizeSql(String sql) {
-    if (!NORMALIZATION_ENABLED) {
-      return sql;
-    }
-    try {
-      return SqlSanitizer.sanitize(sql);
-    } catch (Exception e) {
-      log.debug("Could not sanitize sql", e);
-      return null;
-    }
-  }
-
   public static SqlStatementInfo sanitizeAndExtractInfo(String sql) {
-    try {
-      return SqlStatementInfoExtractor.extract(sanitizeSql(sql));
-    } catch (ParseException e) {
-      log.debug("Could not extract sql info", e);
-      return new SqlStatementInfo(null, null, null);
+    if (!NORMALIZATION_ENABLED) {
+      return new SqlStatementInfo(sql, null, null);
     }
+    return SqlSanitizer.sanitize(sql);
   }
 }
