+++ javaagent-api/javaagent-api.gradle
@@ -1,5 +1,5 @@
 plugins {
-  id 'com.intershop.gradle.javacc' version '4.0.0'
+  id 'org.xbib.gradle.plugin.jflex' version '1.5.0'
 }
 
 group = 'io.opentelemetry.javaagent'
@@ -10,20 +10,15 @@ apply from: "$rootDir/gradle/publish.gradle"
 project.ext.minimumBranchCoverage = 0.0
 project.ext.minimumInstructionCoverage = 0.0
 
-javacc {
-  configs {
-    sanitizer {
-      inputFile = file('src/main/javacc/SqlSanitizer.jj')
-      packageName = 'io.opentelemetry.javaagent.instrumentation.api.db.sanitizer'
-      outputDir = file("${project.buildDir}/generated/javacc/sql")
-    }
-    statmentInfoExtractor {
-      inputFile = file('src/main/javacc/SqlStatementInfoExtractor.jj')
-      packageName = 'io.opentelemetry.javaagent.instrumentation.api.db.sanitizer'
-      outputDir = file("${project.buildDir}/generated/javacc/sql")
-    }
-  }
+def jflexTargetDir = file"${project.buildDir}/generated/jflex/sql"
+def jflexTask = tasks.create("sqlSanitizerJflex", org.xbib.gradle.plugin.JFlexTask) {
+  group = 'jflex'
+  description = 'Generate SqlSanitizer'
+  source = [file("${project.projectDir}/src/main/jflex/SqlSanitizer.flex")]
+  target = jflexTargetDir
 }
+tasks.compileJava.dependsOn(jflexTask)
+sourceSets.main.java.srcDirs += jflexTargetDir
 
 dependencies {
   api deps.opentelemetryApi
